# crates/orchestrator/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "convex-orchestrator",
    "description": "Self-hosted replacement for Convex Cloud's BigBrain orchestrator.",
    "license": {
      "name": "LicenseRef-FSL-1.1-Apache-2.0",
      "identifier": "LicenseRef-FSL-1.1-Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/teams/create_team": {
      "post": {
        "tags": ["teams"],
        "operationId": "create_team",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlatformCreateTeamArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformTeamResponse"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        }
      }
    },
    "/v1/teams/{team_id}/list_members": {
      "get": {
        "tags": ["teams"],
        "operationId": "list_members",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlatformListTeamMembersResponse"
                }
              }
            }
          },
          "401": {
            "description": ""
          }
        }
      }
    },
...
    "/api/dashboard/teams": {
      "get": {
        "tags": ["dashboard"],
        "operationId": "list_teams",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["dashboard"],
        "operationId": "create_team",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/teams/{team_id}": {
      "post": {
        "tags": ["dashboard"],
        "operationId": "update_team",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        }
      }
    },
    "/api/dashboard/teams/{team_id}/delete": {
      "post": {
        "tags": ["dashboard"],
        "operationId": "delete_team",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        }
      }
    },
    "/api/dashboard/teams/{team_id}/members": {
      "get": {
        "tags": ["dashboard"],
        "operationId": "list_members",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamMember"
                  }
                }
              }
            }
          },
          "403": {
            "description": ""
          }
        }
      }
    },
    "/api/dashboard/teams/{team_id}/remove_member": {
      "post": {
        "tags": ["dashboard"],
        "operationId": "remove_member",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveMemberArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        }
      }
    },
    "/api/dashboard/teams/{team_id}/update_member_role": {
      "post": {
        "tags": ["dashboard"],
        "operationId": "update_member_role",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        }
      }
    },
    "/api/dashboard/teams/{team_id}/get_entitlements": {
      "get": {
        "tags": ["dashboard_stubs"],
        "operationId": "get_entitlements_stub",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "stub: returns unlimited self-hosted entitlements",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StubTeamEntitlementsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/teams/{team_id}/unpause": {
      "post": {
        "tags": ["dashboard_stubs"],
        "operationId": "unpause_deployments",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "stub no-op (no pause/unpause flow in self-hosted)"
          }
        }
      }
    },
    "/api/dashboard/teams/{team_id}/invites": {
      "get": {
        "tags": ["dashboard"],
        "operationId": "list_invites",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["dashboard"],
        "operationId": "create_invite",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            }
          },
          "403": {
            "description": ""
          }
        }
      }
    },
    "/api/dashboard/teams/{team_id}/invites/cancel": {
      "post": {
        "tags": ["dashboard"],
        "operationId": "cancel_invite",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelInvitationArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "caller is not an admin of this team"
          },
          "404": {
            "description": "no such pending invitation on this team"
          }
        }
      }
    },
    "/api/dashboard/invites/{code}/accept": {
      "post": {
        "tags": ["dashboard"],
        "operationId": "accept_invite",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "invitation accepted, caller added to team"
          },
          "404": {
            "description": "invitation not found"
          },
          "409": {
            "description": "invitation already accepted"
          }
        }
      }
    },
    "/api/dashboard/teams/{team_id}/apply_referral_code": {
      "post": {
        "tags": ["dashboard_stubs"],
        "operationId": "apply_referral_code",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "stub no-op (no referral program in self-hosted)"
          }
        }
      }
    },
    "/api/dashboard/teams/{team_id}/referral_state": {
      "get": {
        "tags": ["dashboard_stubs"],
        "operationId": "get_referral_state_stub",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "stub: empty referral state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StubReferralState"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/members": {
      "get": {
        "tags": ["admin"],
        "operationId": "list_members",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMembersResponse"
                }
              }
            }
          },
          "403": {
            "description": "not a super-admin"
          }
        }
      }
    },
...
    "/api/admin/teams": {
      "get": {
        "tags": ["admin"],
        "operationId": "list_teams",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTeamsResponse"
                }
              }
            }
          },
          "403": {
            "description": "not a super-admin"
          }
        }
      },
      "post": {
        "tags": ["admin"],
        "operationId": "create_team",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamArgs"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid name or slug already taken"
          },
          "403": {
            "description": "not a super-admin"
          }
        }
      }
    },
...
    "/api/admin/teams/{team_id}/delete": {
      "post": {
        "tags": ["admin"],
        "operationId": "delete_team",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamActionResponse"
                }
              }
            }
          },
          "403": {
            "description": "not a super-admin"
          },
          "502": {
            "description": "a deployment could not be torn down"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateTeamArgs": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": ["string", "null"],
            "description": "Derived from the name when omitted."
          }
        }
      },
...
      "InvitationArgs": {
        "type": "object",
        "required": ["email", "role"],
        "properties": {
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
...
      "PlatformCreateTeamArgs": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": ["string", "null"]
          }
        }
      },
...
      "PlatformListTeamMembersResponse": {
        "type": "object",
        "required": ["members"],
        "properties": {
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformTeamMember"
            }
          }
        }
      },
...
      "PlatformTeamResponse": {
        "type": "object",
        "required": ["id", "name", "slug"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
      "RemoveMemberArgs": {
        "type": "object",
        "required": ["memberId"],
        "properties": {
          "memberId": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "TeamMember": {
        "type": "object",
        "required": ["id", "email", "role"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": ["string", "null"]
          },
          "role": {
            "type": "string"
          }
        }
      },
      "TeamResponse": {
        "type": "object",
        "required": ["id", "name", "slug"],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "creator": {
            "type": ["integer", "null"],
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "UpdateTeamArgs": {
        "type": "object",
        "properties": {
          "name": {
            "type": ["string", "null"]
          },
          "slug": {
            "type": ["string", "null"]
          }
        }
      },
...
    }
  },
...
}